implemented guests and supplies calculation classes and functionality

+++ wedding--simulator-server/wedding-sim/src/main/java/com/weddingsimulator/weddingsim/model/WeddingCostCalculation.java
@@ -33,8 +33,7 @@ public class WeddingCostCalculation {
 
     private int suppliesConfirmed;
 
-    private double suppliesPayed;
+    private int suppliesPayed;
 
     private int totalSupplies;
-
 }
